test: split task intake blocked cases
